SnakeAI is ready to play!
This is the snake game, ready to be used and logging info/warnings. It is configurable until now, by using logging flags (will be detailed upon release). If any bugs are found, fell free to hit me a DM.

A special thank you to @skwow, which made a super useful YouTube tutorial on snake game for Pygame!