DQN implemented on the snake game!
After a long time of learning, we've finally got it running. Special thanks to @farizrahman4u for his repo, for @simoninithomas for his freecodecamp tutorials on DQN + advancements and a lot of sources which gave me rewards examples in which I'm training my model or ideas for the networks to use:

@chuyangliu (https://github.com/chuyangliu/Snake)
    Reward empty = 0.1
    Reward food = 1
    Reward death = -10

@danielegrattarola (https://github.com/danielegrattarola/deep-q-snake/)
    Reward empty = -0.005
    Reward food = 1
    Reward death = -0.5

The sources for the networks are papers on RL for Atari or similar games. It's combined with I saw in the repos also. Thanks a lot for the opportunity to learn, guys!

Next step:
 - Release of version 1.0 of the DQN;
 - Commit the creation of DDQN, DDDQN, DDDQN + PER;
 - Complete experiments with all the parameters, creating gifs, benchmarks and models ready to be used.