Reworking the get_batch() function and fixing a bug when game_over
The get_batch function was working excellent, but I wasn't being able to implement double dqn in it, due to the need to assign the values that the model guessed to the target network. I'm changing it now to make foundation to these changes.

There was also a bug that the board was almost always one unit to the right and this caused the game to make wrong predictions in some situations. This is fixed modifying the "hack" that was done and bringing the game_over flag to the get_game_data function, so we have empty s' states when the game is over on s, a.