Adding types of point in the map and relative actions
Relative action was inspired in @chuyangliu implementation in https://github.com/chuyangliu/Snake/blob/master/docs/algorithms.md#dqn-solver
It makes the game only use 3 directions, LEFT, FORWARD or RIGHT.
Brought the collision check to the snake module, to use game_over flag inside the game class. This will be pushed to the dqn module.
And the point types are necessary only for better predictions, as not everything is the same type of point in the board.